Refactor Single game in collection

+++ gamevalut-server/src/main/java/met/petar_djordjevic_5594/gamevalut_server/model/customUser/FriendDTO.java
@@ -10,7 +10,7 @@ public record FriendDTO(
         Integer id,
         String username,
         String icon,
-        BigInteger hoursPlayed,
+        Double hoursPlayed,
         String inGame
 ) {
 }
