+++ gamevalut-server/src/main/java/met/petar_djordjevic_5594/gamevalut_server/model/game/GameInUserCollectionDetailsDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public record GameInUserCollectionDetailsDTO(
         Integer id,
-        BigInteger playTime,
+        Double playTime,
         LocalDateTime lastPlayed,
         String title,
         String description,
